+++ packages/component/src/Dictation.js
@@ -21,7 +21,8 @@ const {
   useShouldSpeakIncomingActivity,
   useStopDictate,
   useSubmitSendBox,
-  useUIState
+  useUIState,
+  useContinuousListening
 } = hooks;
 
 const {
@@ -44,6 +45,7 @@ const Dictation = ({ onError }) => {
   const setDictateState = useSetDictateState();
   const stopDictate = useStopDictate();
   const submitSendBox = useSubmitSendBox();
+  const continuousListening = useContinuousListening();
 
   const numSpeakingActivities = useMemo(
     () => activities.filter(({ channelData: { speak } = {} }) => speak).length,
@@ -54,8 +56,10 @@ const Dictation = ({ onError }) => {
     ({ result: { confidence, transcript } = {} }) => {
       if (dictateState === DICTATING || dictateState === STARTING) {
         setDictateInterims([]);
-        setDictateState(IDLE);
-        stopDictate();
+        if (!continuousListening) {
+          setDictateState(IDLE);
+          stopDictate();
+        }
 
         if (transcript) {
           setSendBox(transcript);
@@ -65,6 +69,7 @@ const Dictation = ({ onError }) => {
       }
     },
     [
+      continuousListening,
       dictateState,
       setDictateInterims,
       setDictateState,