+++ packages/core/src/actions/setContinuousListening.ts
@@ -1,0 +1,10 @@
+const SET_CONTINUOUS_LISTENING = 'WEB_CHAT/SET_CONTINUOUS_LISTENING';
+
+export default function setContinuousListening(continuousListening) {
+  return {
+    type: SET_CONTINUOUS_LISTENING,
+    payload: { continuousListening }
+  };
+}
+
+export { SET_CONTINUOUS_LISTENING };